Code to Parse Hotels
This has been tested locally, but not on the database connection (I still can't connect :( ). I ended up making a few changes to the schema for the hotels section, since the website I scraped from didn't provide addresses, but did provide other information. The updated version of the schema is towards the end of the google doc where we had the schema for the 2nd deadline.